CleanyDjango/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-import-export==3.0.1
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.7
xlrd==2.0.1
xlwt==1.3.0